[#3] refactor: Renaming require DTO
